[revert] 変更前に戻す
